Medic Interview Application
This is an NodeJS application which will allow users to get a list of
patients, add, and remove patients.

This application will store the first name, middle name, last name,
gender, date of birth, address, ontario health card number, and any
other generic notes.

This NodeJS app uses mongoDB as the database, mongoose for connecting to
the database, and express (EJS as the template) for the server-page
template.